requirements.txt
azure-core
azure-cosmos
azure-identity
azure-keyvault-secrets
azure-storage-blob
pdfplumber
openai
pyspark
pymongo
tika
tiktoken
unidecode
pinecone-client
python-dotenv
json
pexpect
subprocess32

